rename mpc_checkpoint to checkpoint
