+++ integration-tests/tests/cases/mod.rs
@@ -19,12 +19,12 @@ use test_log::test;
 pub mod canton;
 pub mod canton_stream;
 pub mod chains;
+pub mod checkpoint;
 pub mod compat;
 pub mod ethereum;
 pub mod ethereum_stream;
 pub mod helpers;
 pub mod mpc;
-pub mod mpc_checkpoint;
 pub mod mpc_with_stream;
 pub mod nightly;
 pub mod solana;
